chat-backend/requirements.txt
django>=4.0,<5.0
django-cors-headers
django-webpack-loader
djangorestframework
django-countries
django-extensions
django-solo
requests
typing
uwsgi
pyuwsgi
tqdm
mysql
mysqlclient
coverage
coverage-badge
tzdata
discourse-client-in-disraptor==0.0.8
channels
websocket-client
websockets
ws4py
pyyaml
